requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# OpenAI
openai>=1.3.0

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0

# Video Generation
moviepy==1.0.3
imageio[ffmpeg]>=2.25.0
decorator>=4.4.2
proglog>=0.1.10

# TTS
gtts>=2.4.0

# Image Processing
pillow>=10.0.0

# Utilities
numpy>=1.24.0

langchain-text-splitters>=0.0.1